backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Caching & Queue
redis==5.0.1
hiredis==2.3.2

# WebSocket & Real-time
websockets==12.0
python-socketio==5.11.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

